+++ src/vm/heap/fwd.hpp
@@ -1,7 +1,11 @@
 #ifndef TIRO_VM_HEAP_FWD_HPP
 #define TIRO_VM_HEAP_FWD_HPP
 
+#include "common/defs.hpp"
+
 namespace tiro::vm {
+
+enum class ChunkType : u8;
 
 struct HeapStats;
 struct PageLayout;
