+++ test/unit_tests/vm/heap/heap_test.cpp
@@ -1,6 +1,7 @@
 #include <catch2/catch.hpp>
 
 #include "common/scope_guards.hpp"
+#include "vm/heap/allocator.hpp"
 #include "vm/heap/heap.hpp"
 #include "vm/heap/memory.hpp"
